# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/users/{userId}/role-grants": {
      "post": {
        "operationId": "RoleGrantsController_create",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleGrantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleGrantResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "role-grants"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "RoleGrantsController_findAll",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from a previous page's nextCursor.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleGrantListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "role-grants"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/users/{userId}/role-grants/{roleGrantId}": {
      "delete": {
        "operationId": "RoleGrantsController_revoke",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleGrantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleGrantResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "role-grants"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/schools/{schoolId}/role-grants/invite-candidate": {
      "get": {
        "operationId": "RoleGrantsController_lookupInviteCandidate",
        "parameters": [
          {
            "name": "schoolId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phone",
            "required": false,
            "in": "query",
            "description": "E.164 format",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteCandidateResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "role-grants"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/schools/{schoolId}/instructors": {
      "post": {
        "operationId": "InstructorsController_create",
        "parameters": [
          {
            "name": "schoolId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInstructorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstructorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "instructors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "InstructorsController_findAll",
        "parameters": [
          {
            "name": "schoolId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from a previous page's nextCursor.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstructorListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "instructors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/schools/{schoolId}/instructors/eligible-users": {
      "get": {
        "operationId": "InstructorsController_findEligibleUsers",
        "parameters": [
          {
            "name": "schoolId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EligibleInstructorListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "instructors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/instructors/{id}": {
      "get": {
        "operationId": "InstructorsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstructorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "instructors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "InstructorsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInstructorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstructorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "instructors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "ULTM8 API",
    "description": "Modular REST API (Decision 70).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateRoleGrantDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "INSTRUCTOR",
              "BRANCH_STAFF"
            ]
          },
          "schoolId": {
            "type": "string"
          },
          "branchId": {
            "type": "string",
            "description": "Required when role=BRANCH_STAFF (must belong to schoolId). Optional for INSTRUCTOR — Phase 1's schema.prisma models a School-scoped Instructor grant (branchId null) as the default shape; see the Phase 2 summary for the §6.1 wording this follows rather than re-litigates."
          }
        },
        "required": [
          "role",
          "schoolId"
        ]
      },
      "RoleGrantResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "STUDENT",
              "SCHOOL_OWNER_MANAGER",
              "BRANCH_STAFF",
              "INSTRUCTOR",
              "FRANCHISE_OWNER",
              "GUARDIAN"
            ]
          },
          "userId": {
            "type": "string"
          },
          "userFirstName": {
            "type": "string"
          },
          "userSurname": {
            "type": "string"
          },
          "franchiseId": {
            "type": "string",
            "nullable": true
          },
          "schoolId": {
            "type": "string",
            "nullable": true
          },
          "branchId": {
            "type": "string",
            "nullable": true
          },
          "grantedById": {
            "type": "string",
            "nullable": true
          },
          "grantedAt": {
            "type": "string"
          },
          "revokedAt": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "role",
          "userId",
          "userFirstName",
          "userSurname",
          "grantedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "RoleGrantListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleGrantResponseDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items"
        ]
      },
      "InviteCandidateResponseDto": {
        "type": "object",
        "properties": {
          "found": {
            "type": "boolean"
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "surname": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "found"
        ]
      },
...
      "CreateInstructorDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "The User this profile belongs to. Must already hold an active INSTRUCTOR RoleGrant at this School (matching branchId, if set)."
          },
          "branchId": {
            "type": "string",
            "description": "Branch to scope this profile to. Omit for a School-wide profile."
          },
          "photoUrl": {
            "type": "string"
          },
          "beltRanking": {
            "type": "string",
            "description": "Plain display text (e.g. \"Black Belt, 3rd Dan\") — not a live reference into the grading system."
          },
          "specializations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "phone": {
            "type": "string",
            "description": "School-facing contact number, E.164 — distinct from this User's own login phone."
          },
          "yearsOfExperience": {
            "type": "number",
            "minimum": 0
          },
          "bio": {
            "type": "string"
          }
        },
        "required": [
          "userId"
        ]
      },
      "InstructorResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "schoolId": {
            "type": "string"
          },
          "branchId": {
            "type": "string",
            "nullable": true
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          },
          "beltRanking": {
            "type": "string",
            "nullable": true
          },
          "specializations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "yearsOfExperience": {
            "type": "number",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "schoolId",
          "specializations",
          "createdAt",
          "updatedAt"
        ]
      },
      "InstructorListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InstructorResponseDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "Cursor for the next page, or null if this is the last page."
          }
        },
        "required": [
          "items"
        ]
      },
      "EligibleInstructorUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "firstName",
          "surname",
          "email"
        ]
      },
      "EligibleInstructorListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EligibleInstructorUserDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "UpdateInstructorDto": {
        "type": "object",
        "properties": {
          "specializations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "branchId": {
            "type": "string",
            "description": "Branch to scope this profile to. Pass null to clear (make it School-wide).",
            "nullable": true
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          },
          "beltRanking": {
            "type": "string",
            "description": "Plain display text (e.g. \"Black Belt, 3rd Dan\") — not a live reference into the grading system.",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "description": "School-facing contact number, E.164 — distinct from this User's own login phone. Pass null to clear.",
            "nullable": true
          },
          "yearsOfExperience": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "bio": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}